Feat: NFC-e + DANFE modal + IBPTax + Adquirentes + Contas a Receber + Caixa melhorias
- NFC-e: emissão manual XML layout 4.00, assinatura digital, SOAP SEFAZ PR
- DANFE: modal com iframe (80mm térmico), botões Imprimir/WhatsApp/OK
- IBPTax: entidade, importação CSV, background service vigência, tela Fiscal
- vTotTrib calculado por item via IBPTax (Lei 12.741/2012)
- Adquirentes: CRUD com bandeiras e tarifas por modalidade
- Contas a Receber: auto-geração ao finalizar venda
- Caixa: conferência, pagamento split, integração NFC-e
- Configurações: accordion Fiscal/NFC-e (CSC, série, ambiente, regime)
- Filiais: campo Código IBGE Município
- Help: documentação completa do módulo Fiscal/NFC-e

+++ backend/ZulexPharma.Domain/Entities/VendaItem.cs
@@ -11,7 +11,7 @@ public class VendaItem
     public string ProdutoNome { get; set; } = string.Empty;
     public string? Fabricante { get; set; }
     public decimal PrecoVenda { get; set; }
-    public decimal Quantidade { get; set; } = 1;
+    public int Quantidade { get; set; } = 1;
     public decimal PercentualDesconto { get; set; }
     public decimal PercentualPromocao { get; set; }
     public decimal ValorDesconto { get; set; }
